third_party/blink/web_tests/platform/generic/fast/dom/Element/scrollTop-scrollLeft-strict-quirks-modes-expected.txt

Test for scrollTop/Left values of document.body and document.documentElement in documents in quirks and strict modes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.scrollingElement is document.documentElement
PASS iframeStrict.contentDocument.scrollingElement is iframeStrict.contentDocument.documentElement
PASS iframeStrict.contentDocument.documentElement.scrollTop is 5000
PASS iframeStrict.contentDocument.documentElement.scrollLeft is 4000
PASS iframeStrict.contentDocument.documentElement.scrollHeight is 10020
PASS iframeStrict.contentDocument.documentElement.scrollWidth is 10010
PASS iframeStrict.contentDocument.body.scrollTop is 0
PASS iframeStrict.contentDocument.body.scrollLeft is 0
PASS iframeStrict.contentDocument.body.scrollHeight is 10010
PASS iframeStrict.contentDocument.body.scrollWidth is 10005
PASS iframeStrict.contentWindow.pageYOffset is 5000
PASS iframeStrict.contentWindow.pageXOffset is 4000
PASS iframeQuirks.contentDocument.scrollingElement is iframeQuirks.contentDocument.body
PASS iframeQuirks.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks.contentDocument.documentElement.scrollHeight is 10020
PASS iframeQuirks.contentDocument.documentElement.scrollWidth is 10010
PASS iframeQuirks.contentDocument.body.scrollTop is 5000
PASS iframeQuirks.contentDocument.body.scrollLeft is 4000
PASS iframeQuirks.contentDocument.body.scrollHeight is 10020
PASS iframeQuirks.contentDocument.body.scrollWidth is 10010
PASS iframeQuirks.contentWindow.pageYOffset is 5000
PASS iframeQuirks.contentWindow.pageXOffset is 4000
PASS successfullyParsed is true

TEST COMPLETE
